rotaract-api/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
Faker==33.1.0
fastapi==0.115.6
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysqlclient==2.2.6
numpy==2.2.1
openpyxl==3.1.5
orjson==3.10.12
pandas==2.2.3
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.4
pydantic-extra-types==2.10.1
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.18.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uvicorn==0.34.0
watchfiles==1.0.3
websockets==14.1
